+++ ninja-servlet-jpa-blog-archetype/src/main/resources/archetype-resources/src/main/java/dao/ArticleDao.java
@@ -34,10 +34,15 @@ import com.google.inject.Inject;
 import com.google.inject.Provider;
 import com.google.inject.persist.Transactional;
 
+import org.owasp.html.PolicyFactory;
+import org.owasp.html.Sanitizers;
+
 public class ArticleDao {
    
     @Inject
     Provider<EntityManager> entitiyManagerProvider;
+    
+    public static PolicyFactory sanitizer = Sanitizers.FORMATTING.and(Sanitizers.BLOCKS);
     
     @UnitOfWork
     public ArticlesDto getAllArticles() {
@@ -107,8 +112,9 @@ public class ArticleDao {
         if (user == null) {
             return false;
         }
-        
-        Article article = new Article(user, articleDto.title, articleDto.content);
+        String title  = sanitizer.sanitize(articleDto.title);
+        String content  = sanitizer.sanitize(articleDto.content);
+        Article article = new Article(user, title, content);
         entityManager.persist(article);
         
         return true;